+++ packages/audit/lib/index.ts
@@ -2,4 +2,8 @@ export { Audit } from './audit';
 
 export * from './enums';
 
-export * from './types';
+export * from './interfaces/auditResponse';
+export { AuditParams } from './types/auditParams';
+export { AuditEvent } from './types/auditEvent';
+export { AuditEventPage } from './interfaces/auditEventPage';
+export { AuditEventListResponse } from './interfaces/auditEventListResponse';
