+++ packages/audit/lib/interfaces/auditEventPage.ts
@@ -1,0 +1,6 @@
+export interface AuditEventPage {
+    size: number
+    totalElements: number
+    totalPages: number
+    number: number
+}
