+++ packages/audit/lib/interfaces/auditResponse.ts
@@ -1,0 +1,3 @@
+import { VetchResponse } from '@vonage/vetch';
+
+export type AuditResponse<T> = VetchResponse<T>
