+++ packages/audit/lib/types/auditEvent.ts
@@ -1,0 +1,14 @@
+import { AuditEventTypes } from '../enums';
+
+export type AuditEvent = {
+    id: string
+    eventType: AuditEventTypes
+    createdAt: string
+    userEmail?: string
+    userId?: string
+    accountId: string
+    source: string
+    sourceDescription: string
+    sourceCountry: string
+    context?: object
+}
