+++ packages/audit/lib/types/auditParams.ts
@@ -1,0 +1,10 @@
+import { AuditEventTypes } from '../enums';
+
+export type AuditParams = {
+    eventType?: AuditEventTypes
+    dateFrom?: string
+    dateTo?: string
+    searchText?: string
+    page?: number
+    size?: number
+}
